requirements.txt
astropy==5.2.2
bitsandbytes==0.45.3
diffusers==0.32.2
einops==0.8.1
fil_finder==1.7.2
huggingface_hub==0.29.1
imutils==0.5.4
langchain==0.2.15
langchain_core==0.2.40
langchain_huggingface==0.0.3
langchain_openai==0.1.25
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.11.0.86
pandas==2.0.3
Pillow==10.4.0
plantcv==4.6
pyclipper==1.3.0.post6
pycocotools==2.0.7
python-dotenv==1.0.1
Requests==2.32.3
safetensors==0.5.3
scikit_learn==1.3.2
scipy==1.8.0
Shapely==2.0.7
six==1.17.0
scikit-image
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
git+https://github.com/mobiusml/hqq@306e30d
bitsandbytes>=0.43.1
diffusers>=0.30.0
protobuf==3.20
sentencepiece==0.1.99
